requirements.txt
# Core (lock NumPy to avoid ABI issues)
numpy==1.26.4

setuptools

# Image + OCR basics
pillow==10.3.0
pytesseract==0.3.10

# OpenCV (compatible with NumPy 1.26)
opencv-python==4.6.0.66

# Paddle stack (stable combo)
paddlepaddle==2.6.2
# paddlepaddle-gpu==2.6.2  # only if you actually use CUDA
paddleocr==2.7.0.3

# DocTR (works with Python 3.12 + torch backend)
python-doctr==0.8.1

# EasyOCR (needs torch; this combo works)
easyocr==1.7.1

# ONNX OCR
rapidocr-onnxruntime==1.3.24
onnxruntime==1.17.3

# Torch (required by doctr + easyocr)
torch==2.2.2
torchvision==0.17.2

transformers
sentencepiece
